Added datafiles for weapons,armor, and weapons. Also added subtype filter for non-weapons
