+++ components/RunewordListItem.js
@@ -5,6 +5,9 @@ export default function RunewordListItem({ key, runeword }) {
       <span className={styles.name}>{runeword.name}</span>
       <span className={styles.type}>{runeword.type}</span>
       <span className={styles.level}>min level: {runeword.minLevel}</span>
+      {runeword.subType ? (
+        <span className={styles.subType}>subType: {runeword.subType}</span>
+      ) : null}
     </li>
   )
 }
