+++ components/RunewordListItem.module.css
@@ -16,3 +16,7 @@
     margin: 2px 2px;
 
 }
+.subType {
+    font-style: italic;
+    margin: 2px 2px;
+}
