+++ data/armor.js
@@ -1,0 +1,9 @@
+const ARMOR_RUNEWORDS = [
+  {
+    name: 'Enigma',
+    type: 'Armor',
+    minLevel: 30,
+  },
+]
+
+export default ARMOR_RUNEWORDS
