+++ data/index.js
@@ -1,0 +1,10 @@
+import ARMOR_RUNEWORDS from './armor'
+import SHIELDS_RUNEWORDS from './shields'
+import WEAPONS_RUNEWORDS from './weapons'
+
+const ALL_RUNEWORDS = [
+  ...ARMOR_RUNEWORDS,
+  ...SHIELDS_RUNEWORDS,
+  ...WEAPONS_RUNEWORDS,
+]
+export default ALL_RUNEWORDS
