--- data/runewords.js
@@ -1,19 +1,0 @@
-const ALL_RUNEWORDS = [
-  {
-    name: 'Enigma',
-    type: 'Armor',
-    minLevel: 30,
-  },
-  {
-    name: 'Dodge',
-    type: 'Armor',
-    minLevel: 10,
-  },
-  {
-    name: 'Shield',
-    type: 'Shield',
-    minLevel: 17,
-  },
-]
-
-export default ALL_RUNEWORDS
