+++ data/shields.js
@@ -1,0 +1,9 @@
+const SHIELDS_RUNEWORDS = [
+  {
+    name: 'Dodge',
+    type: 'Shield',
+    minLevel: 30,
+  },
+]
+
+export default SHIELDS_RUNEWORDS
