+++ data/weapons.js
@@ -1,0 +1,10 @@
+const WEAPONS_RUNEWORDS = [
+  {
+    name: 'Breath of the Dying',
+    type: 'Weapon',
+    subType: 'Sword',
+    minLevel: 50,
+  },
+]
+
+export default WEAPONS_RUNEWORDS
